requirements.txt
numpy
pandas
biopython
tqdm
seaborn
matplotlib
scipy
line_profiler